refactor(models): add status_payment field to Transaction model's fillable
This commit updates the 'Transaction' model (app/Models/Transaction.php) by adding the 'status_payment' field to the 'fillable' array.

The 'status_payment' field has been added to the list of mass-assignable attributes, allowing this field to be set during mass assignment, such as when creating or updating a transaction record.

This change is in alignment with the recent addition of the 'status_payment' column to the 'transactions' table in the database, and it ensures that the field can be conveniently assigned when working with transaction data.